+++ src/main/java/com/moeasy/moeasy/config/SecurityConfig.java
@@ -1,7 +1,8 @@
 package com.moeasy.moeasy.config;
 
 
-import com.moeasy.moeasy.jwt.JwtAuthenticationFilter;
+import com.moeasy.moeasy.config.jwt.CustomAuthenticationEntryPoint;
+import com.moeasy.moeasy.config.jwt.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -14,7 +15,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import com.moeasy.moeasy.jwt.CustomAuthenticationEntryPoint;
 
 
 @Configuration
@@ -22,39 +22,40 @@ import com.moeasy.moeasy.jwt.CustomAuthenticationEntryPoint;
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-    private final CustomAuthenticationEntryPoint customAuthenticationEntryPoint;
+  private final JwtAuthenticationFilter jwtAuthenticationFilter;
+  private final CustomAuthenticationEntryPoint customAuthenticationEntryPoint;
 
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(AbstractHttpConfigurer::disable)
-                .cors(Customizer.withDefaults())
-                .exceptionHandling(exception -> exception
-                        .authenticationEntryPoint(customAuthenticationEntryPoint)) // 예외 처리 핸들러 등록
-                .authorizeHttpRequests(authorize -> authorize
-                        // 로그인, 회원가입 등 인증이 필요 없는 경로는 permitAll()로 설정
-                        .requestMatchers(
-                                "/survey",
-                                "/account/test/token",
-                                "/account/login",
-                                "/questions/verifyQrCode",
-                                "/account/callback",
-                                "/moiz/v3/api-docs/**",
-                                "/moiz/swagger/**",
-                                "/moiz/swagger-ui/**").permitAll()
-                        // 그 외 모든 요청은 인증 필요
-                        .anyRequest().authenticated()
-                )
-                // 세션을 사용하지 않으므로 STATELESS로 설정
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
+  @Bean
+  public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    http
+        .csrf(AbstractHttpConfigurer::disable)
+        .cors(Customizer.withDefaults())
+        .exceptionHandling(exception -> exception
+            .authenticationEntryPoint(customAuthenticationEntryPoint)) // 예외 처리 핸들러 등록
+        .authorizeHttpRequests(authorize -> authorize
+            // 로그인, 회원가입 등 인증이 필요 없는 경로는 permitAll()로 설정
+            .requestMatchers(
+                "/survey",
+                "/account/test/token",
+                "/account/login",
+                "/questions/verifyQrCode",
+                "/account/callback",
+                "/moiz/v3/api-docs/**",
+                "/moiz/swagger/**",
+                "/moiz/swagger-ui/**").permitAll()
+            // 그 외 모든 요청은 인증 필요
+            .anyRequest().authenticated()
+        )
+        // 세션을 사용하지 않으므로 STATELESS로 설정
+        .sessionManagement(
+            session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+        .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
 
-        return http.build();
-    }
+    return http.build();
+  }
 
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+  @Bean
+  public PasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
 }
