+++ src/main/java/com/moeasy/moeasy/config/response/enums/CommonErrorCode.java
@@ -1,0 +1,16 @@
+package com.moeasy.moeasy.config.response.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum CommonErrorCode implements ErrorCode {
+  INVALID_PARAMETER(HttpStatus.BAD_REQUEST, "Invalid parameter included"),
+  RESOURCE_NOT_FOUND(HttpStatus.NOT_FOUND, "Resource not exists"),
+  INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "Internal server error");
+
+  private final HttpStatus httpStatus;
+  private final String message;
+}
