+++ src/main/java/com/moeasy/moeasy/config/response/enums/ErrorCode.java
@@ -1,0 +1,12 @@
+package com.moeasy.moeasy.config.response.enums;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+
+  String name();
+
+  HttpStatus getHttpStatus();
+
+  String getMessage();
+}
