+++ src/main/java/com/moeasy/moeasy/config/response/enums/UserErrorCode.java
@@ -1,0 +1,15 @@
+package com.moeasy.moeasy.config.response.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum UserErrorCode implements ErrorCode {
+
+  INACTIVE_USER(HttpStatus.FORBIDDEN, "User is inactive");
+
+  private final HttpStatus httpStatus;
+  private final String message;
+}
