+++ src/main/java/com/moeasy/moeasy/config/swagger/SwaggerConfig.java
@@ -1,0 +1,40 @@
+package com.moeasy.moeasy.config.swagger;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+  @Bean
+  public OpenAPI openAPI() {
+    Info info = new Info()
+        .title("Moeasy API")
+        .version("v1.0.0")
+        .description("Moeasy API 문서");
+
+    // Security Scheme 이름
+    String jwtSchemeName = "jwtAuth";
+    // API 요청 헤더에 인증 정보 포함
+    SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwtSchemeName);
+    // SecuritySchemes 등록
+    Components components = new Components()
+        .addSecuritySchemes(jwtSchemeName, new SecurityScheme()
+            .name(jwtSchemeName)
+            .type(SecurityScheme.Type.HTTP) // HTTP 방식
+            .scheme("bearer")
+            .bearerFormat("JWT")); // 토큰 형식 지정
+
+    return new OpenAPI()
+        .info(info)
+        .addSecurityItem(securityRequirement)
+        .components(components);
+
+
+  }
+}
