+++ src/main/java/com/moeasy/moeasy/config/swagger/SwaggerExamples.java
@@ -1,4 +1,4 @@
-package com.moeasy.moeasy.response.swagger;
+package com.moeasy.moeasy.config.swagger;
 
 public class SwaggerExamples {
