--- src/main/java/com/moeasy/moeasy/jwt/JwtAuthenticationFilter.java
@@ -1,89 +1,0 @@
-package com.moeasy.moeasy.jwt;
-
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.moeasy.moeasy.response.FailApiResponseDto;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.security.SignatureException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.lang.NonNull;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtUtil jwtUtil;
-    private final UserDetailsService userDetailsService;
-    private final ObjectMapper objectMapper; // ObjectMapper 주입
-
-    @Override
-    protected void doFilterInternal(
-            @NonNull HttpServletRequest request,
-            @NonNull HttpServletResponse response,
-            @NonNull FilterChain filterChain
-    ) throws ServletException, IOException {
-
-        final String authHeader = request.getHeader("Authorization");
-
-        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        try {
-            final String jwt = authHeader.substring(7);
-            final String userEmail = jwtUtil.extractEmail(jwt);
-
-            if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-                UserDetails userDetails = this.userDetailsService.loadUserByUsername(userEmail);
-
-                if (jwtUtil.validateToken(jwt, userDetails.getUsername())) {
-                    UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
-                            userDetails,
-                            null,
-                            userDetails.getAuthorities()
-                    );
-                    authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                    SecurityContextHolder.getContext().setAuthentication(authToken);
-                }
-            }
-        } catch (ExpiredJwtException e) {
-            setErrorResponse(response, HttpStatus.UNAUTHORIZED, "Token has expired. Please log in again.");
-            return;
-        } catch (SignatureException | MalformedJwtException e) {
-            setErrorResponse(response, HttpStatus.UNAUTHORIZED, "Invalid token signature or format.");
-            return;
-        } catch (Exception e) {
-            logger.error("An unhandled exception occurred in JWT filter: "+ e.getMessage());
-            setErrorResponse(response, HttpStatus.INTERNAL_SERVER_ERROR, "Server error during authentication.");
-            return;
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private void setErrorResponse(HttpServletResponse response, HttpStatus status, String message) throws IOException {
-        response.setStatus(status.value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-
-        FailApiResponseDto failResponse = FailApiResponseDto.fail(status.value(), message);
-        response.getWriter().write(objectMapper.writeValueAsString(failResponse));
-    }
-}
