+++ src/main/java/com/moeasy/moeasy/service/aws/S3Manager.java
@@ -1,0 +1,39 @@
+package com.moeasy.moeasy.service.aws;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.presigner.S3Presigner;
+
+@Configuration
+public class S3Manager {
+
+  @Value("${cloud.aws.credentials.access-key}")
+  private String accessKey;
+  @Value("${cloud.aws.credentials.secret-key}")
+  private String secretKey;
+  @Value("${cloud.aws.region.static}")
+  private String region;
+
+  @Bean
+  public S3Client amazonS3Client() {
+    AwsBasicCredentials credentials = AwsBasicCredentials.create(accessKey, secretKey);
+    return S3Client.builder()
+        .region(Region.of(region))
+        .credentialsProvider(StaticCredentialsProvider.create(credentials))
+        .build();
+  }
+
+  @Bean
+  public S3Presigner s3Presigner() {
+    AwsBasicCredentials credentials = AwsBasicCredentials.create(accessKey, secretKey);
+    return S3Presigner.builder()
+        .region(Region.of(region))
+        .credentialsProvider(StaticCredentialsProvider.create(credentials))
+        .build();
+  }
+}
