+++ src/main/java/com/moeasy/moeasy/service/question/MakeQuestionService.java
@@ -3,6 +3,7 @@ package com.moeasy.moeasy.service.question;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.moeasy.moeasy.config.response.custom.CustomFailException;
 import com.moeasy.moeasy.dto.quesiton.MultipleChoiceIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.MultipleChoiceQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.OnboardingMakeQuestionRequestDto;
@@ -10,7 +11,6 @@ import com.moeasy.moeasy.dto.quesiton.OnboardingQuestionAnswerDto;
 import com.moeasy.moeasy.dto.quesiton.QuestionResponseDto;
 import com.moeasy.moeasy.dto.quesiton.ShortAnswerIncludeIdQuestionDto;
 import com.moeasy.moeasy.dto.quesiton.ShortAnswerQuestionDto;
-import com.moeasy.moeasy.response.custom.CustomFailException;
 import com.moeasy.moeasy.service.llm.NaverCloudStudioService;
 import java.util.ArrayList;
 import java.util.Arrays;
